chore: move normalize.py to utils/language_norm and convert tests to doctests
