+++ API_REFERENCES.md
@@ -52,6 +52,8 @@
   * [validate\_input](#yasbd.utils.input_validator.validate_input)
 * [yasbd.utils.language\_classifier](#yasbd.utils.language_classifier)
   * [classify\_language](#yasbd.utils.language_classifier.classify_language)
+* [yasbd.utils.language\_normalizer](#yasbd.utils.language_normalizer)
+  * [normalize\_lang](#yasbd.utils.language_normalizer.normalize_lang)
 * [yasbd.utils.logger](#yasbd.utils.logger)
   * [log\_info](#yasbd.utils.logger.log_info)
 * [yasbd.utils.paragraph\_stream](#yasbd.utils.paragraph_stream)
@@ -696,6 +698,66 @@ selected candidates.
 **Raises**:
 
 - `ValueError` - If the detector returns no language scores.
+
+<a id="yasbd.utils.language_normalizer"></a>
+
+# yasbd.utils.language\_normalizer
+
+<a id="yasbd.utils.language_normalizer.normalize_lang"></a>
+
+#### normalize\_lang
+
+```python
+@validate_input
+def normalize_lang(lang_code: str) -> str
+```
+
+Normalize a language tag to an ISO-639-1 language code.
+
+The helper is explicit and opt-in. It does not alter the core
+``BoundaryDetector`` language handling.
+
+**Examples**:
+
+  
+  >>> normalize_lang("EN")
+  'en'
+  >>> normalize_lang("en-US")
+  'en'
+  >>> normalize_lang("en-Latn")
+  'en'
+  >>> normalize_lang("pt-BR")
+  'pt'
+  >>> normalize_lang("")
+  ''
+  >>> normalize_lang("   ")
+  ''
+  >>> normalize_lang("not-a-language-code")  # doctest: +ELLIPSIS
+  Traceback (most recent call last):
+  ...
+- `yasbd.exceptions.InvalidInputError` - ...
+  >>> normalize_lang("akk")  # doctest: +ELLIPSIS
+  Traceback (most recent call last):
+  ...
+- `yasbd.exceptions.InvalidInputError` - ...
+  
+
+**Arguments**:
+
+- `lang_code` - A language code or tag, such as ``"EN"``, ``"en-US"``,
+  or ``"en-Latn"``.
+  
+
+**Returns**:
+
+  A two-letter ISO-639-1 language code, or an empty string if empty input.
+  
+
+**Raises**:
+
+- `ImportError` - If the optional ``langcodes`` dependency is missing.
+- `InvalidInputError` - If the tag cannot be parsed or does not resolve to a
+  two-letter ISO-639-1 language code.
 
 <a id="yasbd.utils.logger"></a>
